Refactor: Member 엔티티 코드 리펙터링 및 디렉토리 이동
- 'email', 'nickname', 'phone_number' 필드에 unique = true 추가
- createdAt' 필드에 updatable = false 추가
- 'createdAt' 필드에 '@CreationTimestamp' 어노테이션 추가
- 'emailVerified'와 'otpVerified' 필드의 데이터 타입을 Boolean에서 boolean으로 변경
- 'emailVerified'와 'otpVerified' 필드에 'ColumnDefault("false")' 및 'columnDefinition = "BOOLEAN DEFAULT false"' 추가
- 'role' 필드의 'ENUM' 컬럼 정의에 명시적 값 추가
