--- src/main/java/org/example/spring/domain/Member.java
@@ -1,66 +1,0 @@
-package org.example.spring.domain;
-
-import jakarta.persistence.Column;
-import jakarta.persistence.Entity;
-import jakarta.persistence.Enumerated;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-import java.time.Instant;
-import lombok.AccessLevel;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import org.example.spring.constants.Gender;
-import org.hibernate.annotations.ColumnDefault;
-
-@Builder
-@AllArgsConstructor(access = AccessLevel.PRIVATE)
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-@Getter
-@Entity
-public class Member {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "user_id", nullable = false)
-    private Long id;
-
-    @Column(name = "email", nullable = false, length = 100)
-    private String email;
-
-    @Column(name = "password", nullable = false)
-    private String password;
-
-    @Column(name = "nickname", nullable = false, length = 100)
-    private String nickname;
-
-    @Column(name = "name", nullable = false, length = 100)
-    private String name;
-
-    @Column(name = "phone_number", nullable = false, length = 100)
-    private String phoneNumber;
-
-    @Enumerated
-    @Column(name = "gender", nullable = false)
-    private Gender gender;
-
-    @Column(name = "created_at", nullable = false)
-    private Instant createdAt;
-
-    @Column(name = "updated_at")
-    private Instant updatedAt;
-
-    @Column(name = "last_login_date", nullable = false)
-    private Instant lastLoginDate;
-
-    @ColumnDefault("false")
-    @Column(name = "email_verified", nullable = false)
-    private Boolean emailVerified;
-
-    @ColumnDefault("false")
-    @Column(name = "otp_verified", nullable = false)
-    private Boolean otpVerified;
-
-}
