+++ src/main/java/org/example/spring/domain/member/Member.java
@@ -1,0 +1,72 @@
+package org.example.spring.domain.member;
+
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.EnumType;
+import jakarta.persistence.Enumerated;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import java.sql.Timestamp;
+import lombok.AccessLevel;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import org.example.spring.constants.Gender;
+import org.hibernate.annotations.ColumnDefault;
+import org.hibernate.annotations.CreationTimestamp;
+
+@Builder
+@AllArgsConstructor(access = AccessLevel.PRIVATE)
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+@Getter
+@Entity
+public class Member {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    @Column(name = "member_id", nullable = false)
+    private Long id;
+
+    @Column(name = "email", nullable = false, unique = true, length = 100)
+    private String email;
+
+    @Column(name = "password", nullable = false, length = 68)
+    private String password;
+
+    @Column(name = "nickname", nullable = false, unique = true, length = 100)
+    private String nickname;
+
+    @Column(name = "name", nullable = false, length = 100)
+    private String name;
+
+    @Column(name = "phone_number", nullable = false, unique = true, length = 20)
+    private String phoneNumber;
+
+    @Enumerated(EnumType.STRING)
+    @Column(name = "gender", nullable = false)
+    private Gender gender;
+
+    @CreationTimestamp
+    @Column(name = "created_at", updatable = false,nullable = false)
+    private Timestamp createdAt;
+
+    @Column(name = "updated_at")
+    private Timestamp updatedAt;
+
+    @Column(name = "last_login_date")
+    private Timestamp lastLoginDate;
+
+    @ColumnDefault("false")
+    @Column(name = "email_verified", nullable = false, columnDefinition = "BOOLEAN DEFAULT false")
+    private boolean emailVerified;
+
+    @ColumnDefault("false")
+    @Column(name = "otp_verified", nullable = false, columnDefinition = "BOOLEAN DEFAULT false")
+    private boolean otpVerified;
+
+    @Enumerated(EnumType.STRING)
+    @Column(name = "role", nullable = false, columnDefinition = "ENUM('USER', 'ADMIN', 'BANNED') DEFAULT 'USER'")
+    private MemberRole role;
+}
